Remove ResourceController
